Problem: managing development and deployment experience of Omnigres apps
It's still far away from the imagined ideal. In a few projects, we were able to
use a some pretty lean Makefiles to orchestrate most of work, but it is not ideal
and they were not very refined.

Solution: implement specialized command line tooling

We could have further improved those Makefiles, but they are less useful when
it comes to adoption – they'd work but would be not as reliable or flexible,
nor would they be as convenient as specialized tooling.

This initial source code snapshot is in no way "final" – an initial prototype.