Problem: maintaining source-of-truth schema
This is generally not done as incremental migrations have been widely
adopted. However, the tide has been changing on this – the better we can
automated figuring out a correct, non-lossy path from schema A to schema
B, the easier our lives will be.

Solution: integrate (very primitively) with migra

While I don't think migra will be ultimate solution to this, it provides
conceptual support for showing the diff between versions of the schema.

This change brings this problem domain into scope. We'll iterate on it
and see where we land. Lots of opportunities.

Since Migra is being operated out of a container, we are now putting
Omnigres containers into their own private networks. For now, every
container gets its own network.