Problem: Run/Start output can be frustrating since it is long and does not give a clear idea of the progress when image has to be downloaded. This will avoid frustrating users by showing the progress bar tracking downloaded docker layers.
The progress is tracked per layer adding the the in-flight downloads progress.
We display a final message at the end to reassure the user the CLI is not stuck, since in big images the last percentage might take a while.
An error message is shown quitting the progress bar with ESC to avoid
allowing the execution to break on the Docker SDK (since it won't be
able to find the image).