Problem: terminating `omnigres run` leaves container running
Reported in https://github.com/omnigres/cli/pull/4

When I press Ctrl-C, the container stays. This was not the case before
and it was intended to be cleaned up.

Solution: restore the behavior gone in 0e8ee39d78b1c35cc235169583df8a1ed2485d02

This ensures that we forcibly stop the container if there was an error
or the container was attached to.