Problem: database flag and orbs arguments were confusing, how can we map multiple orbs to a single database?
Solution: Drop the arguments and database flag and rely only on the
directory until we better understand what is the best way to override
the defaults.
The defaults as implemented here will use the current orb as the first
directory below the workspace on the current directory tree.
When the current workspace is the current directory, or the current
directory is outside the workspace tree then we consider all orbs listed
in the config.