Problem: We cannot migrate running as the user postgres from inside the container since the database server won't have write permissions on the bound source code directory.
Solution: We run as a user without name, using the Uid from the host
system. This works on Linux, still to be tested on other OS.
In order to do so we follow a workaround described on
https://github.com/docker-library/docs/blob/master/postgres/README.md#arbitrary---user-notes

PostgreSQL image must start the deault entrypoint with the user that
owns the data directory.
So we create a volume that is initiated by the postgres user (PostgreSQL
also requires that the user to run initdb exists in /etc/passwd).
Then we run the container with a custom entry point just to change
ownership for the current OS user.
Finally we run again with the default entry point but passing the
current user to the docker container user parameter.